move: 폴더구조 변경, 폴더 이름 단수로 변경, 서비스를  플랫폼 으로 변경 #20
