+++ src/auth/dto/sign-in.dto.ts
@@ -1,4 +1,4 @@
 import { PickType } from '@nestjs/swagger';
-import { Users } from 'src/users/entities/users.entity';
+import { Users } from 'src/user/entities/users.entity';
 
 export class SignInDto extends PickType(Users, ['email', 'password']) {}
