+++ src/categorie/categories.controller.ts
@@ -3,7 +3,7 @@ import { CategoriesService } from './categories.service';
 // import { CreateCategoryDto } from './dto/create-category.dto';
 // import { UpdateCategoryDto } from './dto/update-category.dto';
 
-@Controller('categories')
+@Controller('categorie')
 export class CategoriesController {
   constructor(private readonly categoriesService: CategoriesService) {}
