+++ src/notification/notifications.controller.ts
@@ -1,17 +1,8 @@
 
-import {
-  Controller,
-  Get,
-  Post,
-  Body,
-  Patch,
-  Param,
-  Delete,
-} from '@nestjs/common';
 import { Controller, Get, Request } from '@nestjs/common';
 import { NotificationsService } from './notifications.service';
 
-@Controller('notifications')
+@Controller('notification')
 export class NotificationsController {
   constructor(private readonly notificationsService: NotificationsService) {}
