+++ src/platform/entities/service-prices.entity.ts
