+++ src/platform/platforms.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { PlatformsService } from './platforms.service';
+import { PlatformsController } from './platforms.controller';
+
+@Module({
+  controllers: [PlatformsController],
+  providers: [PlatformsService],
+})
+export class PlatformsModule {}
