+++ src/platform/platforms.service.ts
@@ -3,7 +3,7 @@ import { CreateServiceDto } from './dto/create-service.dto';
 import { UpdateServiceDto } from './dto/update-service.dto';
 
 @Injectable()
-export class ServicesService {
+export class PlatformsService {
   create(createServiceDto: CreateServiceDto) {
     return 'This action adds a new service';
   }
