--- src/services/services.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ServicesService } from './services.service';
-import { ServicesController } from './services.controller';
-
-@Module({
-  controllers: [ServicesController],
-  providers: [ServicesService],
-})
-export class ServicesModule {}
