+++ src/user/users.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Get, Req } from '@nestjs/common';
 import { UsersService } from './users.service';
 
-@Controller('users')
+@Controller('user')
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
